fix(mcp): isolate hpc backend workers
PicklingError: Can't pickle run_mace_singleArguments was triggered by dill recursing into the worker's module globals and hitting a FastMCP-generated Pydantic class. Fix by isolating backend worker functions in modules with no FastMCP state so Parsl can serialize import-clean workers for MACE, XANES, and gRASPA. hpc_misc_mcp.py was inspected and left unchanged because it has no backend worker callables.
