+++ src/chemgraph/mcp/mace_mcp_hpc.py
@@ -27,12 +27,13 @@ from chemgraph.execution.utils import (
     resolve_structure_files,
 )
 from chemgraph.mcp.cg_fastmcp import CGFastMCP
+from chemgraph.mcp.mace_worker import _ls_remote_files, _mace_worker
 from chemgraph.mcp.transfer_tools import register_transfer_tools
 from chemgraph.schemas.mace_parsl_schema import (
     mace_input_schema,
     mace_input_schema_ensemble,
 )
-from chemgraph.tools.parsl_tools import extract_output_json, run_mace_core
+from chemgraph.tools.parsl_tools import extract_output_json
 
 logger = logging.getLogger(__name__)
 
@@ -71,70 +72,6 @@ mcp = CGFastMCP(
           name -- do not pass it.
     """,
 )
-
-
-# ── Worker (runs on the backend) ───────────────────────────────────────
-
-
-def _mace_worker(job: dict) -> dict:
-    """Execute a single MACE simulation on a backend worker.
-
-    Accepts a *job dict* (not the schema) so the pre-submit hook can
-    attach transport keys ``inline_structure`` / ``remote_structure_file``
-    before submission.
-    """
-    import json
-    import tempfile
-
-    job = dict(job)
-
-    # Pre-staged remote file: use the path directly on the worker FS.
-    remote_file = job.pop("remote_structure_file", None)
-    if remote_file is not None:
-        job["input_structure_file"] = remote_file
-        if not os.path.isabs(job.get("output_result_file", "")):
-            job["output_result_file"] = os.path.join(
-                os.path.dirname(remote_file),
-                job.get("output_result_file", "output.json"),
-            )
-
-    # Inline structure: materialise on the worker's filesystem.
-    inline = job.pop("inline_structure", None)
-    if inline is not None:
-        from ase import Atoms
-        from ase.io import write as ase_write
-
-        atoms = Atoms(
-            numbers=inline["numbers"],
-            positions=inline["positions"],
-            cell=inline.get("cell"),
-            pbc=inline.get("pbc"),
-        )
-        tmpdir = tempfile.mkdtemp(prefix="chemgraph_mace_")
-        xyz_path = os.path.join(tmpdir, "structure.xyz")
-        ase_write(xyz_path, atoms)
-        job["input_structure_file"] = xyz_path
-        if not os.path.isabs(job.get("output_result_file", "")):
-            job["output_result_file"] = os.path.join(
-                tmpdir, job.get("output_result_file", "output.json")
-            )
-
-    output_file = job.get("output_result_file")
-    if output_file:
-        os.makedirs(os.path.dirname(os.path.abspath(output_file)), exist_ok=True)
-
-    params = mace_input_schema(**job)
-    result = run_mace_core(params)
-
-    # When inline, embed full output so the caller doesn't need to read
-    # a file on the remote filesystem to recover the results.
-    if inline is not None and isinstance(result, dict):
-        out_file = job.get("output_result_file", "")
-        if os.path.isfile(out_file):
-            with open(out_file) as fh:
-                result["full_output"] = json.load(fh)
-
-    return result
 
 
 # ── Pre-submit transport hook ──────────────────────────────────────────
@@ -212,13 +149,6 @@ def run_mace_single(params: mace_input_schema) -> dict:
 
 
 # ── Ensemble fanout ────────────────────────────────────────────────────
-
-
-def _ls_remote_files(path: str) -> list[str]:
-    """Backend-side helper: list non-directory entries in *path*."""
-    return sorted(
-        f for f in os.listdir(path) if os.path.isfile(os.path.join(path, f))
-    )
 
 
 def _expand_mace_ensemble(params: mace_input_schema_ensemble) -> list[dict]:
